[INTERNAL] Demo Kit: Samples which can be opened in new window are with default theme
Problem: Samples are opened with the configuration mentioned in their html
file. This might result in opening the sample with different than the
default theme (Quartz Light).

Solution: We make sure we add the correct search param at the end of the
sample's URL, in order to force the default theme.

Change-Id: I0dc0a35675bd342a4313f4632c3ea22c391ed95d
BCP: 2180098702