[FIX] sap.uxap.ObjectPageLayout: Unset selected section of hidden page
Problem:
The qunit for "unset selected section of hidden page" was failing
intermittently when it hit the following condition:

When a hidden page is made visible, two resize listeners trigger:
1) listener for resize of its *content*
2) listener for resize of its *content container*
and in the timing conditions of the test (a previously hidden page is
made freshly visible) the changes triggered by the two listeners were
in conflict:
--the first listener handles changes in the overflow(see overflow-anchor css property)
  => it updated the "selectedSection" to match the changed scroll position
--the second listener ensures that the required scroll of a previously hidden page is applied
  => it updated the scroll position to match the specified "selectedSection"

=> the two actions contradict

Solution:
Ensured that the action of the first (*content* resize) listener is only
scoped to the cases that it is meant to address. Because in this case (the case
of showing a previously hidden page) we do not have a real content rezise but a
change from invisible to visible.

Change-Id: Id07f2ad1dff064a6db38b5fec5ccec1a987c6d6a
BCP: 2180046524
(cherry picked from commit 16333552192568f43bd14648bf70041bf666e4de)
CR-Id: 002075125800001052492021