[INTERNAL] sap.ui.fl - fl delegate extended by delegate types
In the past it was possible to define only complete (with read and
write functions) delegates as default delegate. We decided to offer
the possibility to split the fl delegates into a read and a writeonly
parts. Complete delegates are also still possible. We are able now
to define more then one default delegate for a specific model.
The delegate mediator merges then both parts together if required.

Change-Id: Ibf322502d276aa626b1d0c64f11883409f4b3df4
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
JIRA: TEAM42-89