[FIX] sap.m.DatePicker: referencing invisible text elements as aria labelled by now works
Problem description:
The implementation assumed that only a sap.m.Label
control instances will be referenced in the "aria-labellby"
attribute of the sap.m.DatePicker control DOM representation.
The text for the title of the control dialog on mobile,
was created by using the referenced control DOM representations.

Solution:
Different controls can now be referenced in the "aria-labellby"
attribute of the control DOM representation, which includes
invisible text elements.
The "LabelEnablement" static class is used, in order to get the
referenced controls instead of their DOM representations.

Change-Id: Ic7d3a6842a843d538cb4b0031644adad825ccc77
CR-Id: 002075125900001531102021
BCP: 002075129400000848162021