[INTERNAL] sap.m.MultiComboBox: Token indication class toggled correctly
Previously, when a token has been created via "setSelectedKey()"
method after the initial creation of the control, the class indicating
if there are tokens in the input - "sapMMultiComboBoxHasToken" was not toggled correctly.
This was causing the placeholder text to not be removed after value tokenization and
also, the tokens paddings were not added.

Now the addition/removal of the class is fixed.

BCP: 2180098236
Change-Id: I2c8dc51fb0140729321ac1129375c58906b23015