[FIX] sap.ui.layout.ResponsiveSplitter: No truncation of last area
Problem description:
When a bar is moved its sibling areas are turned to %-based areas by the AssociativeSplitter.
Then when the root (ResponsiveSplitter) got shrunk, those % could exceed the maximum
available space.

Solution:
Mark such %-based areas and shrink them upon parent resizing if possible.

Example:
1000px available width, 1 area with 500px and 1 with 50% (500 + 500 = 1000 OK)
Available width shrunk to 800px, 1 area with 500px and 1 with 50% (500 + 400 > 800 Not OK)

Note: Only % that are set from the AssociativeSplitter are reduced,
not % set by the app developer.

In addition:
- Refactored AssociativeSplitter. Same logic with small differences was implemented
on 2 places. Now it is all in the base class - Splitter
- Improved logging when there is no enough space to shrink anymore (misconfigured panes)
- Added tests

Snippix: 58445
BCP: 2180118515
Change-Id: I2143b46dae75299e666a3836b88a826ccb387202