[FIX] sap.f.DynamicPageTitle: Cutting of buttons is fixed
Problem: When the title is too long, the buttons in the actions
aggregation is getting cut, as this area does not have min-width. In
DynamicPageTitle, this can be fixed, using areaShrinkRatio property.
However, in ObjectPageDynamicHeaderTitle there is no such property.

Solution: We are checking whether there is any visible control in the
OverflowToolbar, placed in the content/actions aggregation and if so -
min-width of 3rem is set.

Change-Id: I1e62bb6fa37ae10470b3169c957eb0710fc0515a
BCP: 2180085793