[INTERNAL] Demo Kit: Demo Kit Sample page change theme issue fixed
Problem:
If we are on Quartz Dark appearance, we can't apply Quartz Light theme
in the Sample page options menu.

Root Cause:
The Side Panel (ControlsMaster) is keeping its own property for the current
theme and it doesn't take into account the change of appearances of the
Demo Kit ("Light", "Dark", "Auto").

Solution:
Every time we save the newly selected options of the Side Panel options
menu, we check what is the current theme and the theme that the user
has inputed. If they are different, then we apply the user's input.

Additionally a slight optimisation of the appearance logic is made.
Also we now fire an event that the theme is changed, every time we
change the appearance.

Change-Id: I524674fed8b4983d92003b97f63c5800e2f539f2
BCP: 2180137546