[FIX] sap.ui.documentation: fix timing and environment issues
When the SDK was executed in an environment with unexpected neo-app.json
or versionoverview.json content, it could run into errors. The
evaluating code has been made more robust.

There was a race condition between calculation of web page title texts
(based on the sap-ui-version.json) and the routing. The code has been
made more robust.

When a page could not be found, the route handler called setTimoeut with
an object, not a function. This resulted in an eval() call, causing a
SyntaxError.

Change-Id: I08c74ed3621da7238df9c6a988d36a236dffa373