[FIX] XMLTemplateProcessor: hide stashed objects only where supported
Previously, for stashed objects, the XMLTP automatically has set
'visible' to false. When the target class did not support the 'visible'
property, this caused assertion failures in ManagedObject.

With this change, the 'visible' property is only set to false when the
target class supports that property and when the property is of type
boolean.

Additionally, handling of namespaced attributes has been fixed
(Element.prototype.setAttributeNS requires a qualified attribute name).

Change-Id: I1145423ea02dcdff2ae2db9e4a7547646124a54e