[INTERNAL] ColumnListItem: fix Delete & Detail button margin-left with Dummy column rendered
There is a margin-left: 0.125rem applied to the ListItemBase on the Detail
& Delete buttons of the item.
This margin works as expected where there is no dummy column present. Since the
<tr> element has padding with, the above buttons do not overflow the item.
But when dummy column is present, the padding of the list item is not applied.
Due to this the border of the dummy column overlaps the button.
Hence to fix the issue of button overflowing, the margin left is removed.
BCP: 2180136458

Change-Id: Ie5582f20f41317909e57dc814c1eb1e162b86ca8