[FIX] sap.m.Table: strict layout single column handling bug is fixed
 - For non strict layout sap.m.Table always expect at least one column
with "auto" width. If there is only a single column in the table,
without looking at the width property of the Column, the Table renders
that column as it has width "auto" but this logic contradicts with the
strict layout(fixedLayout=Strict).
 - In Strict layout every column must be as wide as defined in the API.
 - Single column handling is removed for the strict layout.

 - pointer-events: none is added to the pseudo-elements of custom focus
handling of the table which was causing twice click to navigate.

 - Growing button alignment with dummy column issue is fixed.

 - Bottom border of the dummy column must be shown at the dummy cell of
the last row but if the growing button is not visible. If growing button
is visible border must be shown next to the growing button.

 - Custom focus handling of the dummy column had focus outline issues in
HCB and HCW themes and those are fixed.

 - Unit tests are added

BCP: 2180127458
BCP: 2180136742
BCP: 2180129886
BCP: 2180129839

Change-Id: I27c10d7a4ad91e495b341277d9130e6c2664ff46