[FIX] sap.m.Table: Doucble click to navigate issue is fixed
 - pointer-events: none is added to the pseudo-elements of custom focus
handling of the table which was causing twice click to navigate.

 - Growing button alignment with dummy column issue is fixed.

 - Bottom border of the dummy column must be shown at the dummy cell of
the last row but if the growing button is not visible. If growing button
is visible border must be shown next to the growing button.

 - Custom focus handling of the dummy column had focus outline issues in
HCB and HCW themes and those are fixed.

 - Unit tests are adapted

Change-Id: I27c10d7a4ad91e495b341277d9130e6c2664ff46
(cherry picked from commit 09d12d2b52b0ba16ca578accacb92b9ab95b0067)
CR-Id: 002075125800000102042021
BCP: 2180136742
BCP: 2180129886
BCP: 2180129839