[INTERNAL] mdc.Table: DataStateIndicator integration
 - PluginBase is now designed to forward plugin owner for composite
controls:
To forward the plugin owner, the composite control can implement
"get[PluginName]PluginOwner" method. This method can be used to return
the internal control instance which will be responsible for the plugin.
In this case even though the plugin instance is added to the composite
control, the return value of "get[PluginName]PluginOwner" will be the
logical owner of the plugin.
If composite control instantiate internal controls asynchronous then
"get[PluginName]PluginOwner" method can return a promise. Afterwards
when the promised is resolved "get[PluginName]PluginOwner" method must
return the instance of the internal control that will be logical owner
of the plugin.

 - Private messageLinkVisible property is removed. Setting messageLink
to empty text should be enough to make the link invisible. This behavior
is now aligned with the showMessage method.

 - enableFiltering public property is added. By setting this property to
true, the DataStateIndicator enables filtering for data state messages,
and the user is shown a link where the data state messages can be
filtered. This feature encapsulates the filtering behavior in the
plugin. This will help us to reuse this implementation in SmartTable,
mdc.Table and standalone tables. To allow hook into filtering
capabilities, applyFilter and clearFilter public events are added which
allow preventing the default behavior. applyFilter event triggered with
filter object and revert function which can be used to clear filters
from the consumer side. In addition a private filterInfoPress event can
be used to make the generated info toolbar active and the event handler
gets triggered when the info toolbar is pressed. This can be used e.g.
by the SmartTable to open p13n Dialog when the info toolbar is pressed.

 - DSI#refresh method is now able to refresh filtered messages.

 - Unit Tests are added and adapted

Jira: CPOUIFTEAMB-1670

Change-Id: I4350b7956fdd5be1f8cdb8542e24011437d9a3f4