[INTERNAL][FIX] Table: Row heights not updated after flip animation
The issue happens in a specific use case with NavContainer with 2 Pages, each containing a ui.Table
and navigation between the pages with Flip animation. The second table detaches the 'transitionend'
event for both tables, therefore the row heights are not updated when navigated back to the first
Table.
The fix prevents this by adding the table id as an event namespace and thereby removing the event
handler only for the correct table.

Change-Id: I2050e8a0c8a81ef082246c901ee207dd0c06f59e
BCP: 002075129400001419982021
(cherry picked from commit 7f00d4208fb2c9be79d5a1150e5ee5bf64013c7a)
CR-Id: 002075125800000256642021