[FIX] Revert "[INTERNAL] core.routing: propagate hash change to component only when browser hash contains its part"
This reverts commit 77f1935df4ef6c0acb23e424aebb22807c4f096d.

Reason for revert: when a router is still active, a hashchange event still needs to be fired to the router when there's no hash for this router in the browser hash. Furthermore, a router will be restarted after processing the hashchange event from its parent, it's not possible to check to which router the hashchange should be forwarded.

Change-Id: I0b7e2477704b6786a50c1c4d310ccace7481d944
BCP: 2170014800