[FIX] FilterField: Filter only for InParameters in ConditionModel
There might be InParameter that point to data outside the ConditionModel
and cannot be used as Filter.
To indicate the InParameters pointing to the same ConditionModel, the
"conditions" part is stored in the name of the InParameter inside the
condition. Only such InParameters are used as filter. 

To support old variants the InParameters are transformed while
importing.

Change-Id: I440945266f1af1970a0bbc6ffa714325b2604a13
BCP: 2180101445