[INTERNAL] sap.m.Select: fix announcment of new selection
Problem: New value is no longer read out by screen readers when picker is closed
and arrow keys are used to change the selecton.

Solution: An input[type=button] element is now used for the focused element
to properly announce the new selection when popover is closed and
value is changed with arrows.

Change-Id: I79ab5c9bfbe5e8aa06f2897d483fb8870e054c93
BCP: 2180103686
BCP: 2170038887
BCP: 2170040571
BCP: 2170002330