[INTERNAL] Table: fix popinChanged event
The internal _aHiddenInPopin was always updated when the condition in the
onAfterRendering was not met. Due to this the popinChanged event was not fired,
when initially _aHiddenInPopin.length == 0 and later it become > 0, since the else
part always updated in the internal _aHiddenInPopin and the length mismatch did
not fire the popinChanged event.
To fix the unnecessary updates of the _aHiddenInPopin, the onAfterRendering will
only update the _aHiddenInPopin when its undefined. Later when the length of the popins
change, the if condition should update internal _aHiddenInPopin and subsequently
fire the popinChanged event.
BCP: 2180137057

Change-Id: Ia62fbac5f52d02605ca6fb653072dabff51abe1b
(cherry picked from commit 1f53ea64ac9e9ea731279dd4712e21df5db46cd9)
CR-Id: 002075125800000255412021