[INTERNAL] sap.f.IllustratedMessage: Fixed UX and VD feedback
After the initial merge of the IllustratedMessage in the master, there
were a few UX and VD reported recommendations. This is the change fixes
them. They are explained in detail in the linked JIRA.

Additionally a small private feature was implemented. Now it's not
possible to have "widow" words (singe words on the last row) in the title
or the description.

Change-Id: Ib13d7417afaad430b5493e6ff5ebdf24073b10d6
JIRA: BGSOFUIPIRIN-5095