[INTERNAL] StandardListItem: fix adaptTitleSize with wrapping feature
When adaptTitleSize=true, there is fixed height applied to the DIV within
the StandardListItem. This fixed height truncated the title to be visible
when wrapping feature is enabled for items having both title and description.
The fixed height is to be assigned to the DIV only when there is no description
text. This is now fixed.
BCP: 2180137368

Change-Id: I97cf65e6e85ee04b762e2bdb29830e0726f86335