[INTERNAL] sap.ui.ux3: Remove unused CSS rule
This has been identified as the referenced background image doesn't exist and also
seems to never have existed at all.

sapUiUx3TICollectionThingAction is not used in any renderer or other less/css files.
It seems that the rule has been added with the initial control implementation
(not part of OpenUI5 history), but the corresponding renderer code was not active
and has been removed but the rule was still kept.

Therefore the rule can be removed without any side effects.

Change-Id: I96de56604cdd4bf85c0639f436e456311a606ee8