[FIX] sap.m.ObjectNumber: Labelling enabled
This change introduces an ariaLabelledBy association
to the sap.m.ObjectNumber, allowing labelling via
sap.m.Label or direct references in ariaLabelledBy.

If labelled, in order for the label(s) to not overwrite
ObjectNumber's content, relevant information such as
number, unit, emphasizing and state is now referenced in
aria-labelledby as well.

Change-Id: I3aa8241a4ae610e8260465e339b3fba60eab490e
BCP: 2180115276