[FEATURE] UIComponent: support async createContent implementation
When implementing the IAsyncContentCreation interface a UIComponent
can have a async createContent implementation. If the components 
createContent implementation returns a view (created with
the legacy view factory) we wait for the view to be loaded. 
In addition we switch to the new complete async factory for the 
rootView creation as well for nested views in XMLViews, XMLFragments
and HTMLFragments.

Change-Id: Icdd79a9832acaf98fba9769699b5f09dbf9d78a8