[INTERNAL] NavContainer (JSDoc): fix inconsistent case of event names
In the runtime as well as in the event documentation of interface
'sap.m.NavContainerChild', the events 'BeforeHide', 'AfterHide',
'BeforeFirstShow', 'BeforeShow' and 'AfterShow' all are written with the
first letter in upper case.

But in other documentation, references to those events had been written
with the first letter in lower case.

With this change, the references have been aligned with the spelling in
the original event documentation, which also matches the event names
used at runtime.

Change-Id: I2709d93ab4739d983a66a4d1e1eff29b7ea8d5f6