[INTERNAL] ColumnResizer: standalone table support and double click to auto resize column
1. Column resizing capabilites are now enabled with the standalone
sap.m.Table control use case.
The column headers are made focusable and relevant style class are added.
2. Double clicking on the column resizer now enables auto resizing of the column
If a column is to be decreased then the oSession.iMaxDescrease is taken into
consideration to ensure that the column is not reduced more than the
max decrease limit.
Similarly, if a column is to be increaed than the oSession.iMaxIncrease is
taken into consideration to ensure that the column is not increased more than
the max increase limit.
JIRA: CPOUIFTEAMB-1765
JIRA: CPOUIFTEAMB-1764

Change-Id: Ifb3728ce229d37f5b9f2269ca59a2943bf621bf0