[FIX][INTERNAL] sap.ui.fl...Loader does not use the cache key while reInitialize
In case an application is changed within the session and a reinitialize is triggered, the cache key provided at the startup cannot be used.
This commit ensures, the cache key is not passed in this scenario.

Change-Id: I46e526cddc13b26d6bbd3f1022fe5312b92ba2dc
CR-Id: 002075125800000531582021
BCP: 2170012946
(cherry picked from commit 9ea49028f45a7c9ef59e8a6eae0043d24680249a)