[INTERNAL]Engine: modularize UI creation
The Engine is no longer handling the creation of the
outer Popup (E.g. Dialog, Popover creation). The UIManager
is now handling the creation of popups while the Engine
is only implementing the core functionality for adaptation.

The UIManager now also allows to open several p13n UIs in one
P13nWrapper, this Wrapper is currently not being used and only
in place for the basic functionality and will be refined in an
additional change.

Change-Id: I1ed911605947741a1c298127736190dce4d02960