[FIX] sap.m.OverflowToolbar: Missing buttons with bound LayoutData
Problem: In some cases, when the OverflowToolbarLayoutData is bound and
there is a change of its "priority" property just after the rendering,
the LayoutData is not getting invalidated and does not fire a change
event to its parent. This way, the OverflowToolbar works with the old
value of the "priority" property and some buttons are missing - they are
not shown in the toolbar, neither in the Popover.

Solution: Always make sure that the OverflowToolbarLayoutData will be
invalidated and the change event will be fired to the OverflowToolbar,
when "priority" property is bound and the invalidation is suppressed.

BCP: 2080179140
Change-Id: Ie695ebe72f73afce300e403a0ee2317e5a2d4149