[FIX] sap.m.SegmentedButton: Items' accessibility enhanced
If a user has icon-only item with a tooltip, then that tooltip
should be placed in aria-label as well, so that the control
can generate its accessible name with that text, instead of
the default name of the icon.

In addition to this, for cases where user hasn't provided an
alternative text via a tooltip, icon-only items would try to get
an icon's descriptive text first and only fall back to its name
as a last resort.

BCP: 2170017803

Change-Id: I61fe1c04ba85356860794f5a8b643eeebdc0711a