[INTERNAL] Table: bug fixes
1. The popinChange event was fired based on the changes only in the hiddenInPopin.
This behavior is incorrect and it should be fired on any changes to the popin area.
A new parameter 'visibleInPopin' is now added to the popinChanged event, which provides
all the visible popins in the table. This will help the event listeners to
correctly react to the changes to the popin-area (e.g. SmartTable)
2. The columns are not moved to the popin area, if the contextualWidth of the
table does not change more that 16px. This is done to ensure that in case the
vertical scrollbar is rendered by the parent contains than the columns should
not jump in and out of the popin area. More details are provided as code comment.
3. Since the dummy column is the last column in the table DOM structure, the
style class to make the last internal column of the table to 2.5rem was not
having effect anymore. The Navigation column becomes smaller when the table
is switched to 'Strict' layout as a result of column resizing which also affect
the column resizing functionality making the dummy area later in case a column
is made smaller in the table. This unexpected behavior is now fixed by adapting
the styles and also increasing the iEmptySpace threshold to 3 to avoid increasing
the dummy area due to sub-pixel rendering issue in Chrome.
BCP: 2170038943
BCP: 2170019608
BCP: 2170038183

Change-Id: I26612a53bdeab7d025bcfa4b18a98a3f61efb4e1