[INTERNAL] JSDoc/TypeScript: introduce @ts-skip to omit APIs from d.ts
The new (experimental) JSDoc tag '@ts-skip' acts as a marker for the
generation of TypeScript definitions to skip an API (to omit it from the
generated d.ts files).

It is mainly intended to prevent non-fixable inconsistencies in the
existing UI5 APIs from showing up as compile time errors in TypeScript.

Its usage is for now limited to fields and methods.

Change-Id: I98b30909dff09a7a617e0ebb73205a25e3920fb8