[INTERNAL] ui5loader: improve error reporting
Errors during Dependency Resolution
-----------------------------------
When an error occurs during dependency resolution, then the message of
the outermost error includes the dependency chain that lead to the
innermost error.

Syntax Errors in Sync Mode
--------------------------
When a syntax error occurs during eval() of a synchronously loaded
module, then the location of the error was not revealed (as eval does
not provide such information). The loader already triggered a second
load via script tag (to pinpoint the error location) when running in
debug mode, but this was a rarely known feature.

A new error log entry now provides a hint reg. debug mode and the
alternative loading via script tag is now limited to syntax errors as it
is not helpful / caused confusing error messages for other error types.

Additionally, not the external resource is loaded via script tag, but
the internal script source (which potentially has been transformed, e.g.
during client side coverage). This requires the use of a Blob and might
conflict with a strict Content Security Policy (CSP). But as sync
loading itself already conflicts with such a CSP, the loader's CSP
compliance is not getting worse and error analysis is (hopefully)
simplified.

Other Fixes
-----------
- errors that occur while loading shim dependencies are no longer
  stifled
- when an erroneous module is required again, only the original error
  is reported, not a 'negative cache error'. The 'negative cache error'
  wording often lead to confusion.
- error handling in the context of sync XHRs has been cleaned up:
  When send() fails, the XHR does not provide a meaningful status
- use the error 'name' of the root cause to communicate between
  loader and Core.js. This avoids propagating the loadError flag
  and is more in line with the discussed ES feature 'error cause'
  ( https://github.com/tc39/proposal-error-cause )

Change-Id: Ie9e4c14bd0d4a5d66cb45f5d5b3d57366d67364f