[INTERNAL] sap.ui.fl CompVariantState updates variant names
To update variant names, the runtime instance and the text of a variant has to be separated. This is done by the name property (setName / storeName) similar to favorite, executeOnSelection and contexts.
For having the update enabled, the consuming control has to change from the getText("variantName") to the getName() function of the returned variants.

Change-Id: I8cb4115f42079a888bb8e7dafb1406bd9b13e5fa
JIRA: CLOUDIOI-178