[INTERNAL] Table: Fix QUnit tests that fails if executed on remote system
Depending on whether UI5 is running locally, remotely, or remotely with debug
sources, the modules seem to be loaded differently (in a different order?).

This QUnit test did require an unnecessary test delegate. This test delegate is
not just unnecessary for the test, but also seem to be implemented incorrectly,
because it overwrite some APIs on productive delegates.
But because of this, the tests run fine locally. If the modules are (probably)
loaded in a different order, the APIs are not overwritten, and the test fails.

The problem is solved by implementing a test delegate specifically for this
QUnit module, inside the module.

Change-Id: I940a55235a636395660e28c9e1abd8a5d31b1ef8