[FIX] sap.m.Input: Horizontal padding is now added to the test samples
After horizontal margins were added to the suggestions popup when in
100% width the tabular suggestions started to appear cut off if the
input is not in a container providing some space on both sides.

Since the case where the input is used directly in a page with no side
space was discussed with the UX experts and it was considered a
non-valid use case, now all the test samples of sap.m.Input are fixed
to be in a padding providing container and not stuck to the start
and the end of the browser window.

BCP: 2180113811
Change-Id: I1174cac2864274ce78ef1f39ebc4a8f5c99ce6ac