[FIX] sap.m.GenericTag: Misaligment of text and unit is fixed
Problem: The text and unit of the ObjectNumber are misaligned, due to
some recent changes in the ObjectNumber.

Solution: In GenericTag's styles there is an override for padding-top of
the ObjectNumber's unit, which seems to not be needed anymore. It is now
deleted, so the alignment of text and unit are according to the VD
requirements.

BCP: N/A
Change-Id: I4f7b41b869eefa81144a6e2793071e3a43e1692b