[INTERNAL] DemoKit: Corrected link parsing
Problem:
When the Documentation section is opened from a deep link
that contains a "_" symbol in its hash, the master-side tree
is not expanded to file location

Root cause:
To show the master side we have to parse the file name from the file URI.
If the URI contained a "_" symbol in its hash, the file name could not be
parsed because of incomplete regular expression.

Solution:
Corrected the regular expression to make provision for URIs with
"_" symbol in the hash.

Change-Id: I7d6f269f46bb58f9a2f4cea7ec2a8e8b33dc7d2c
BCP: 2180121035