[INTERNAL] CSP: sap/ui/thirdparty requiresTopLevelScope=false config
Adds requiresTopLevelScope="false" configuration for sap/ui/thirdparty
modules that are exporting globals.
This configuration allows the bundler to wrap the code within an IIFE
to prevent the need of 'eval' when executing the module as part of a
bundle.
The bundler also adds code to assign/export the globals to the
window/this context.

Note: requiresTopLevelScope can only work if the module does not modify
the declared top level variables after the initial execution.
This needs to be checked manually before adding the configuration for a
module.

JIRA: CPOUI5FOUNDATION-232
Change-Id: I253dfbb9268795c8d8d6f35959ef7a849aa033d9