[INTERNAL][FIX] sap.ui.rta Fix reload in case of perso + dirty changes
When entering RTA with perso changes and creating dirty changes on a
clean app, reload behavior was buggy, because the reloadCurrentApp FLP
API call does not update the parsedHash correctly when reloading. Max
Layer parameter will not get removed correctly in that case and and
error will be thrown.

BCP: 2170034511
Change-Id: Ic352bafb9ec5e09f25cb6ab0cdd76d465e615882