[FIX] sap.uxap.ObjectPageLayout: Correct scrollToSection issue
Problem:
When the user presses an anchorBar button to scroll to a section,
and the app simultaneously changes the sections structure (removes/adds blocks)
=> the scroll ends at an incorrect position

Solution:
Remember the chosen section (as selectedSection) prior to scrolling
=> allow the page adjust its final scroll position to match the remembered section
(existing implementation in adjustLayoutAndUXRules).

Notes:
The same solution is already applied if the app called instead "setSelectedSection"
because "setSelectedSection" already remembers the chosen section (as selectedSection)
prior to scrolling => the problem does not occur there.

Change-Id: Ie4e08de98757e5bd748b33ce025225047ca37649
BCP: 2170051076
CR-Id: 002075125800000972032021