[FIX] sap.uxap.ObjectPageSection: Expand option for all hidden sections
Background:
If a section (or any of its subsections) is given importance lower than
'high' it is hidden on smaller screen sizes. In those cases, an "Expand"
and "Show More" buttons are renedred aside the section title to allow
show the hidden content

Problem:
By UX rule, the title header of the first visible section is hidden
=> the "Expand" and "Show More" buttons remained hidden

Solution:
Extend the check for title header visibility to make provision for
section importance.

Change-Id: Ifd0a9411e28537977faaab54bd185c2c679b4b91
BCP: 2170061627