[INTERNAL] sap.m.Select: Update picker ariaLabelledBy association
When an element is referecing an id of an element which is nested in,
screen readers are facing issues and results to redundant speech output.

Problem: When valueState is set, selected value and value state text are read out twice upon opening the picker.

Solution: Set the pickers ariaLabelledBy association to point to an InvisibleText instance
rendered in the static area.

Additionally, an issue with double announcement of the value state is fixed by
removing the aria-describedby from the SelectRenderer.js as at this version the
ValueStateMessage delegate handles the announcment of the value state.
The removed code was unnecessary downported with I30ac54f4272ab937e1141fd33ab55a5d97365bd1.

Change-Id: I82a40e54a80d1c2c9b8fccd88d801b1102b125ea
BCP: 2080361690
CR-Id: 002075125800003516792022
(cherry picked from commit 80475e740a7613b2d2a447f088234f87d1b23fc6)