[FIX] sap.f.DynamicPage: Not clickable area of title is fixed
Problem: When in DynamicPageTitle there is a control, which occupies the
whole available space (VBox, HBox), there are some empty-looking places,
that are not clickable. However, the cursor style is pointer, which
makes these areas look clickable.

Solution: Adjust the CSS styles of the VBox/HBox, placed in the
DynamicPageTitle, so that they do not look clickable (default cursor is
used).

BCP: 2070272080
Change-Id: Ic3c1765b638dce786a03922b8d25b0da80d6eea0