[INTERNAL] Table: Strict Layout - Special Handling for "last" column
If the last remaining column is larger than the table in
fixedLayout=Strict, then this column width applied to the column is
removed thus occupying the available space in the table.
In case the column is smaller than the table then the width is kept and the remaining space is occupied by the dummy column.
JIRA: CPOUIFTEAMB-1787
BCP: 2170110013

Change-Id: I97524ca9feea45deabdf4e01390de9b4c6461955
(cherry picked from commit 48b11cfef5d7a8755b89a1326b6643063438c850)
CR-Id: 002075125800001268612021