[FIX] DemoKit: Restricted UL styling
Problem:
We got feedback from UA that the bullets of unordered list-items within
a datatable should be visible in the general case.

Solution:
Restricted the overwrriding style that was added earlier to apply only
to the "sl" context, as in the reference documentation.

BCP: 2170045100
Change-Id: Id3b478777129c16bebdf36e8eb9af1d483d80ac6
(cherry picked from commit eecc1f2029405886b4a617132f1955fd5738696e)
CR-Id: 002075125800001062982021