[FIX] ManagedObject: fix cloning of associations to root object
When an association in the cloned subtree pointed to the root object,
the association was not adapted in the clone (it still pointed
to the original object).

Additionally, the test for the cloning of associations within nested
binding templates was fixed. It always failed in the first run and only
succeeded in later runs due to QUnit's 'reorder' functionality.

Change-Id: I78bb182c4af465bc94bc6c301258933b688043d1