[FIX] sap.m.Image: Failing QUnit on FireFox is fixed
Problem: QUnit is failing beacuse after placing a Control in the HTML,
we need to call applyChanges, in order the control to be rendered. As in
this case, the control is not rendered, we cannot sroll to it and the
onLoad event of the Image is not fired.

Solution: Call the applyChanges method, so that the Image is rendered
and it's onLoad event is fired.

BCP: 2170094028
Change-Id: I6c8474681989aa386f18feeec32c5a3217c82881