[INTERNAL] sap.ui.fl: fix unstable test '.../extensionPoint/Processor' - 2
in the tests some content of an async view was not available in time,
so the tests wait for the loaded event from that view.
What was overlooked in the last commit was that the async view is also
embedded in the sync view, this is now also properly handled.

BCP: 2170103354
Change-Id: I4efcfdfcc32dab45089b12b72939370d8a8dccba