[INTERNAL] sap.ui.fl SmartVariantManagementWriteAPI offers a override for standardVariants
Overrides allows to set the executeOnSelection flag within the standard
variant as well as reapplies the execution.
This is done in cases where the standard variant is overridden after the
loadVariants call was done (i.e. in case of an outer control sets the
standard variant execute on selection of an inner
SmartVariantManagement-control)

Change-Id: I9e4570761107785b5b883fb6cc28a988da66318c
JIRA: CLOUDIOI-178