[FIX] sap.m.StepInput: on mobile value is set after spin release
Problem:
If a spin is about to end we set the value on mouseup, but we
reset the spin on press. On desktop the events are fired in
this order: mouseup, touchend, press. On mobile the order
is different: touchend, press, mouseup.
Solution:
Now for mobile, we also set the value on touchend, using
the value before the spin is reset.

Change-Id: I4d49b62fe2a823dee91422127509ec579060188d
BCP: 2170076330