[FIX] sap.uxap.ObjectPageLayout: Correct contentLength rounding
Problem:
The value of calculated content length exceeds the actual one with 1px
because it is ceiled.

Solution:
Replace Math.ceil with Math.round

Note:
Math.ceil was added with If3fceaf738c7c4910f560c4fc2d462e6ddaaffc8 to
overcome an IE specific issue. but we no longer support IE in this version

Also corrected the amount of bottom space we add in the case where
we need to keep the current scrollTop but detect that the page scrollHeight
has being changed to a smaller scrollHeight that is unsufficient to
keep the required scrollTop.

Change-Id: Ia4a5c5d36b14501e4fa08070955f0c3adf524056