[INTERNAL] sap.f.IllustratedMessage: Added session timeout sample
All applications that require user sign-in should employ a mechanism
for automating user sign-out. Automatic sign-outs are a result of
inactivity over a predefined period of time. Depending on the security
level of the information presented, the period of inactivity may vary
between 5 and 60 minutes. When the inactivity is about to reach that
threshold, a session timeout dialog appears automatically to inform the
user.

The sample implements this specification by using the sap.m.Dialog and
sap.f.IllustratedMessage.

Change-Id: I67d4af1e3a0d20429e41b3ac2164b0e6a0e88383
JIRA: BGSOFUIPIRIN-5171