[INTERNAL] Table: fix _configureAutoPopin spy qunit test
The test to check if _configureAutoPopin is called when the importance property
of the column header changed failed since the column already had "High"
importance. The importance were applied randomly to the columns and there
can be chances that the column for which the importance is changed already
has High importance. This ramdomization of importance is removed to stabilize
the qunit test.

Change-Id: I7f425b7cabe1965e37e3acb8eb2afd59e4a759e5