[INTERNAL] Demo Kit: Fixed API Reference scroll behaviour
Problem: When we make the entity on the Detail page of the API
Reference to be on full screen and we start scrolling down, the
full screen is turned off.

Root Cause:
When we scroll and match a topic, we always turned off the full screen
even if the user explicitly wanted the topic to be on full screen.
This is a side effect of showing the 'Master' when we are opening the
API Reference from deep link.

Solution:
We are now forcing the show of the 'Master' only initially when we are
coming from a deep link, or the first time we go to an entity.
Now when we scroll we are not forcing the turn off of the full screen.

Additionally a bug is fixed where if we are on a topic in full screen
mode, then we go to a different tab ('Sample'), and we go back, the
'Master' is shown (as it should), but the 'toggleFullScreenBtn' is
displayed in the incorrect state.

Change-Id: Ieb21aaf842b03b3b2210db7d26ab4c0ebbd98fb8
JIRA: BGSOFUIPIRIN-5079